Struct fixed_circular_buffer::CircularBuffer

pub struct CircularBuffer<T> {
    // some fields omitted
}

Fixed size circular/cyclic/ring buffer

A FIFO (first in, first out) queue. It cannot represent an empty buffer.

When constructed, the internal list must not be empty, and cannot contain invalid (e.g. uninitialized) elements.

Methods

impl<T> CircularBuffer<T>

fn len(&self) -> usize

Returns the number of elements (before starting to loop around).

fn queue(&mut self, elem: T) -> T

Enqueues (push at beginning) the given element at the beginning of the buffer Dequeues (pop at end) the last element and returns it This keeps the the buffer length

fn set_first(&mut self, index: usize)

Sets the offset for the first element, relative to the currently first element When index is out of range, it loops around

fn get(&self, index: usize) -> &T

Returns a reference to the element at the given index When index is out of range, it loops around

fn get_mut(&mut self, index: usize) -> &mut T

Returns a mutable reference to the element at the given index When index is out of range, it loops around

fn swap_internal(&mut self, a: usize, b: usize)

Swaps the two elements at the given indices a and b. When a or b are out of range, they loop around

fn swap(&mut self, index: usize, elem: T) -> T

Swaps the element at the given index with the specifiied new one. When a or b are out of range, they loop around

fn iter_circular<'s>(&'s self) -> Skip<Cycle<Iter<'s, T>>>

Returns an iterator over the buffer looping around at the end. This creates a never ending iterator

fn iter<'s>(&'s self) -> Take<Skip<Cycle<Iter<'s, T>>>>

Returns an iterator over the buffer without looping around.

unsafe fn from_raw_parts(list: Box<[T]>, first: usize) -> Self

Constructs the structure from its raw components

Unsafety

This function is unsafe as there is no guarantee that first < list.len(), nor whether list is non-empty.

fn into_raw_parts(self) -> (Box<[T]>, usize)

Deconstructs the structure into its raw components