Struct fixed_circular_buffer::CircularBuffer

pub struct CircularBuffer<T, L = Box<[T]>> { /* fields omitted */ }

Fixed size circular/cyclic/ring buffer

Almost a FIFO (first in, first out) queue. · It cannot represent an empty buffer. · The queue is always filled (Cannot dequeue without queueing).

When constructed, the internal list must not be empty.

Methods

impl<T, L> CircularBuffer<T, L>

pub unsafe fn from_raw_parts(list: L, first: usize) -> Self

Constructs the structure from its raw components.

Unsafety

This function is unsafe as there is no guarantee that first < list.len(), nor whether list is non-empty.

pub fn into_raw_parts(self) -> (L, usize)

Deconstructs the structure into its raw components

impl<T, L> CircularBuffer<T, L> where
L: Deref<Target = [T]>,

pub fn len(&self) -> usize

Returns the number of elements (before starting to loop around).

pub fn iter_circular<'s>(&'s self) -> IterCircular<'s, T>

Returns an iterator over the buffer looping around at the end. This iterator is initially iterating from the most recently queued element to the oldest, and then looping around (relative to queue). This creates a never ending iterator

pub fn iter<'s>(&'s self) -> Iter<'s, T>

Returns an iterator over the buffer without looping around. Iterates from the most recently queued element to the oldest queued element (relative to queue).

pub fn internal_index(&self, index: usize) -> usize

pub fn get(&self, index: usize) -> &T

Returns a reference to the element at the given index. Smaller indices are more recently queued elements (relative to queue) (0 is the newest). When index is out of range, it loops around.

pub fn internal_index_reversed(&self, index: usize) -> usize

impl<T, L> CircularBuffer<T, L> where
L: DerefMut<Target = [T]>,

pub fn swap(&mut self, elem: T) -> T

Swaps the most recently queued element.

pub fn queue(&mut self, elem: T) -> T

Enqueues (push at beginning) the given element at the beginning of the buffer. Dequeues (pop at end) the last element and returns it.

pub fn set_first(&mut self, index: usize)

Sets the offset for the first element, relative to the currently first element. Smaller indices are more recently queued elements (relative to queue) (0 is the newest). When index is out of range, it loops around.

pub fn get_mut(&mut self, index: usize) -> &mut T

Returns a mutable reference to the element at the given index. Smaller indices are more recently queued elements (relative to queue) (0 is the newest). When index is out of range, it loops around.

pub fn swap_internal(&mut self, a: usize, b: usize)

Swaps the two elements at the given indices a and b. Smaller indices are more recently queued elements (relative to queue) (0 is the newest). When a or b are out of range, they loop around.

pub fn swap_at(&mut self, index: usize, elem: T) -> T

Swaps the element at the given index with the specifiied new one. When a or b are out of range, they loop around.

pub fn queue_reversed(&mut self, elem: T) -> T